requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.0.2
gTTS==2.5.4
idna==3.10
mysql-connector-python==9.2.0
mysqlclient==2.2.7
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
